pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ezekit"
version = "0.1.0"
description = "A modular toolkit for dataset preparation — starting with OCR annotation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "IDLE" }]
dependencies = [
    "click>=8.0",
    "fastapi>=0.110",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
ocr = [
    "PyMuPDF>=1.24",
    "Pillow>=10.0",
    "datasets>=2.18",
    "pyarrow>=15.0",
]
auto-detect = [
    "paddleocr>=2.7",
    "paddlepaddle>=2.6",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
ezekit = "ezekit.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ezekit.modules.ocr" = ["ui/**/*"]
